Implementa autenticação para gerenciamento de tarefas por usuario

+++ app/Http/Controllers/TarefaController.php
@@ -6,6 +6,7 @@ use App\Models\Tarefa;
 use Illuminate\Http\Request;
 
 use function Laravel\Prompts\alert;
+use Illuminate\Support\Facades\Auth;
 
 class TarefaController extends Controller
 {
@@ -13,7 +14,7 @@ class TarefaController extends Controller
     public function index()
     {
 
-        $tarefas = Tarefa::all();
+        $tarefas = Auth::user()->tarefas;
         return view('tarefas.index', compact('tarefas'));
     }
     public function create()
@@ -30,7 +31,11 @@ class TarefaController extends Controller
             'prioridade' => 'required|in:baixa,media,alta',
         ]);
 
-        Tarefa::create($request->all());
+        $user = Auth::user();
+
+        $tarefa = new Tarefa($request->all());
+        $tarefa->user_id = $user->id;
+        $tarefa->save();
         return redirect()->route('tarefas.index')->with('success', 'Tarefa criada com sucesso!');
     }
 
@@ -51,6 +56,10 @@ class TarefaController extends Controller
 
         $tarefa = Tarefa::findOrFail($id);
 
+        if ($tarefa->user_id !== Auth::id()) {
+            return redirect()->route('tarefas.index')->with('danger', 'Você não tem permissão para editar esta tarefa.');
+        }
+
         $tarefa->update($request->all());
 
         return redirect()->route('tarefas.index')->with('success', 'Tarefa atualizada com sucesso!');
@@ -58,15 +67,21 @@ class TarefaController extends Controller
 
     public function destroy(Tarefa $tarefa)
     {
+        if ($tarefa->user_id !== Auth::id()) {
+            return redirect()->route('tarefas.index')->with('danger', 'Você não tem permissão para excluir esta tarefa.');
+        }
+
         $tarefa->delete();
         return redirect()->route('tarefas.index')->with('danger', 'Tarefa excluída com sucesso!');
     }
 
     public function concluir($id)
     {
-
         $tarefa = Tarefa::findOrFail($id);
 
+        if ($tarefa->user_id !== Auth::id()) {
+            return redirect()->route('tarefas.index')->with('danger', 'Você não tem permissão para concluir esta tarefa.');
+        }
 
         if ($tarefa->status == 'concluída') {
             return redirect()->route('tarefas.index')->with('warning', 'A tarefa já está concluída!');
@@ -81,7 +96,10 @@ class TarefaController extends Controller
         $query = $request->input('search');
         $prioridade = $request->input('prioridade');
 
-        $tarefas = Tarefa::where('titulo', 'like', "%{$query}%");
+        $usuarioId = Auth::id();
+
+        $tarefas = Tarefa::where('user_id', $usuarioId)
+                     ->where('titulo', 'like', "%{$query}%");
 
         if ($prioridade) {
             $tarefas = $tarefas->where('prioridade', $prioridade);
